compiler/clib: Flush Input() before first read through stdin
This will remove command line arguments from Input() before something
is read through stdin. This should fix the tests added in previous commit.

git-svn-id: https://svn.aros.org/svn/aros/trunk@47799 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/__fdesc.h
@@ -22,9 +22,9 @@ typedef struct _fcb
 } fcb;
 
 /* privflags */
-#define _FCB_ISDIR        1<<0
-#define _FCB_DONTCLOSE_FH 1<<1
-#define _FCB_FLUSHONREAD  1<<2
+#define _FCB_ISDIR        ((unsigned int)1<<0)
+#define _FCB_DONTCLOSE_FH ((unsigned int)1<<1)
+#define _FCB_FLUSHONREAD  ((unsigned int)1<<2)
 
 
 /* file descriptor structure - one per descriptor */
